feat: export JsonFormat and rename format.ts to json-format.ts
